write to string to avoid partial martials
amusing bug - couldn't figure out why it worked locally and not
as an installed gem. Turns out the paths to my installed gems
are longer, and the marshals were getting partly written in one
context but not the other. Solution is to just put them in a string,
and only dump them if they work.